+++ src/lib.rs
@@ -75,6 +75,7 @@ extern crate tokio_io;
 extern crate tokio_executor;
 extern crate tokio_reactor;
 extern crate tokio_threadpool;
+extern crate tokio_tcp;
 
 #[macro_use]
 extern crate log;
@@ -189,20 +190,4 @@ pub mod prelude {
         Poll,
         task,
     };
-}
-
-#[cfg(feature = "unstable-futures")]
-fn lift_async<T>(old: futures::Async<T>) -> futures2::Async<T> {
-    match old {
-        futures::Async::Ready(x) => futures2::Async::Ready(x),
-        futures::Async::NotReady => futures2::Async::Pending,
-    }
-}
-
-#[cfg(feature = "unstable-futures")]
-fn lower_async<T>(new: futures2::Async<T>) -> futures::Async<T> {
-    match new {
-        futures2::Async::Ready(x) => futures::Async::Ready(x),
-        futures2::Async::Pending => futures::Async::NotReady,
-    }
 }
