+++ src/net/mod.rs
@@ -36,9 +36,8 @@
 //! [`UdpFramed`]: struct.UdpFramed.html
 //! [`framed`]: struct.UdpSocket.html#method.framed
 
-mod tcp;
 mod udp;
 
-pub use self::tcp::{TcpStream, ConnectFuture};
-pub use self::tcp::{TcpListener, Incoming};
+pub use tokio_tcp::{TcpStream, ConnectFuture};
+pub use tokio_tcp::{TcpListener, Incoming};
 pub use self::udp::{UdpSocket, UdpFramed, SendDgram, RecvDgram};
