--- src/net/tcp/mod.rs
@@ -1,8 +1,0 @@
-mod incoming;
-mod listener;
-mod stream;
-
-pub use self::incoming::Incoming;
-pub use self::listener::TcpListener;
-pub use self::stream::TcpStream;
-pub use self::stream::ConnectFuture;
