+++ tokio-tcp/README.md
@@ -1,0 +1,15 @@
+# tokio-tcp
+
+TCP bindings for `tokio`.
+
+[Documentation](https://tokio-rs.github.io/tokio/tokio_tcp/)
+
+## License
+
+This project is licensed under the [MIT license](./LICENSE).
+
+### Contribution
+
+Unless you explicitly state otherwise, any contribution intentionally submitted
+for inclusion in Tokio by you, shall be licensed as MIT, without any additional
+terms or conditions.
