+++ tokio-tcp/tests/chain.rs
@@ -1,5 +1,5 @@
 extern crate futures;
-extern crate tokio;
+extern crate tokio_tcp;
 extern crate tokio_io;
 
 use std::net::TcpStream;
@@ -9,7 +9,7 @@ use std::io::{Write, Read};
 use futures::Future;
 use futures::stream::Stream;
 use tokio_io::io::read_to_end;
-use tokio::net::TcpListener;
+use tokio_tcp::TcpListener;
 
 macro_rules! t {
     ($e:expr) => (match $e {
